Trait enum_map::Enum

pub trait Enum: Sized {
    const LENGTH: usize;

    // Required methods
    fn from_usize(value: usize) -> Self;
    fn into_usize(self) -> usize;
}

Enum mapping type.

This trait is implemented by #[derive(Enum)].

This trait is also implemented by bool and u8. While u8 is strictly speaking not an actual enum, there are good reasons to consider it like one, as array of u8 keys is a relatively common pattern.

Required Associated Constants

const LENGTH: usize

Length of the enum.

Required Methods

fn from_usize(value: usize) -> Self

Takes an usize, and returns an element matching into_usize function.

fn into_usize(self) -> usize

Returns an unique identifier for a value within range of 0..Array::LENGTH.

Object Safety

This trait is not object safe.