Crate enum_map

An enum mapping type.

It is implemented using an array type, so using it is as fast as using Rust arrays.

Examples

use enum_map::{enum_map, Enum, EnumMap};

#[derive(Debug, Enum)]
enum Example {
    A,
    B,
    C,
}

let mut map = enum_map! {
    Example::A => 1,
    Example::B => 2,
    Example::C => 3,
};
map[Example::C] = 4;

assert_eq!(map[Example::A], 1);

for (key, &value) in &map {
    println!("{:?} has {} as value.", key, value);
}

Macros

enum_map
Enum map constructor.

Structs

EnumMap
An enum mapping.

IntoIter
A map iterator that moves out of map.

Iter
Immutable enum map iterator

IterMut
Mutable map iterator

Values
An iterator over the values of EnumMap.

ValuesMut
A mutable iterator over the values of EnumMap.

Traits

Enum
Enum mapping type

Derive Macros

Enum
Procedural derive generating enum_map::Enum implementation.