Struct enum_map::EnumMap

pub struct EnumMap<K: Enum<V>, V> { /* fields omitted */ }

An enum mapping.

This internally uses an array which stores a value for each possible enum value. To work, it requires implementation of internal (private, although public due to macro limitations) trait which allows extracting information about an enum, which can be automatically generated using #[derive(Enum)] macro.

Additionally, bool and u8 automatically derives from Enum. While u8 is not technically an enum, it's convenient to consider it like one. In particular, reverse-complement in benchmark game could be using u8 as an enum.

Examples

use enum_map::{enum_map, Enum, EnumMap};

#[derive(Enum)]
enum Example {
    A,
    B,
    C,
}

let mut map = EnumMap::new();
// new initializes map with default values
assert_eq!(map[Example::A], 0);
map[Example::A] = 3;
assert_eq!(map[Example::A], 3);

Implementations

impl<K: Enum<V>, V> EnumMap<K, V>

pub fn values(&self) -> Values<'_, V>

An iterator visiting all values. The iterator type is &V.

Examples

use enum_map::enum_map;

let map = enum_map! { false => 3, true => 4 };
let mut values = map.values();
assert_eq!(values.next(), Some(&3));
assert_eq!(values.next(), Some(&4));
assert_eq!(values.next(), None);

pub fn values_mut(&mut self) -> ValuesMut<'_, V>

An iterator visiting all values mutably. The iterator type is &mut V.

Examples

use enum_map::enum_map;

let mut map = enum_map! { _ => 2 };
for value in map.values_mut() {
    *value += 2;
}
assert_eq!(map[false], 4);
assert_eq!(map[true], 4);

impl<K: Enum<V>, V: Default> EnumMap<K, V>

pub fn new() -> Self

Creates an enum map with default values.

Examples

use enum_map::{Enum, EnumMap};

#[derive(Enum)]
enum Example {
    A,
}

let enum_map = EnumMap::<_, i32>::new();
assert_eq!(enum_map[Example::A], 0);

pub fn clear(&mut self)

Clear enum map with default values.

Examples

use enum_map::{Enum, EnumMap};

#[derive(Enum)]
enum Example {
    A,
    B,
}

let mut enum_map = EnumMap::<_, String>::new();
enum_map[Example::B] = "foo".into();
enum_map.clear();
assert_eq!(enum_map[Example::A], "");
assert_eq!(enum_map[Example::B], "");

impl<K: Enum<V>, V> EnumMap<K, V>

pub fn iter(&self) -> Iter<'_, K, V>

Returns an iterator over enum map.

pub fn iter_mut(&mut self) -> IterMut<'_, K, V>

Returns a mutable iterator over enum map.

pub fn len(&self) -> usize

Returns number of elements in enum map.

pub fn is_empty(&self) -> bool

Returns whether the enum variant set is empty.

This isn't particularly useful, as there is no real reason to use enum map for enums without variants. However, it is provided for consistency with data structures providing len method (and I will admit, to avoid clippy warnings).

Examples

use enum_map::{Enum, EnumMap};

#[derive(Enum)]
enum Void {}

#[derive(Enum)]
enum SingleVariant {
    Variant,
}

assert!(EnumMap::<Void, ()>::new().is_empty());
assert!(!EnumMap::<SingleVariant, ()>::new().is_empty());

pub fn swap(&mut self, a: K, b: K)

Swaps two indexes.

Examples

use enum_map::enum_map;

let mut map = enum_map! { false => 0, true => 1 };
map.swap(false, true);
assert_eq!(map[false], 1);
assert_eq!(map[true], 0);

pub fn as_slice(&self) -> &[V]

Converts an enum map to a slice representing values.

pub fn as_mut_slice(&mut self) -> &mut [V]

Converts a mutable enum map to a mutable slice representing values.

pub fn as_ptr(&self) -> *const V

👎 Deprecated since 0.6.4:
Please use .as_slice().as_ptr() instead

Returns a raw pointer to the enum map's slice.

The caller must ensure that the slice outlives the pointer this function returns, or else it will end up pointing to garbage.

Examples

use enum_map::{enum_map, EnumMap};

let map = enum_map! { 5 => 42, _ => 0 };
assert_eq!(unsafe { *map.as_ptr().offset(5) }, 42);

pub fn as_mut_ptr(&mut self) -> *mut V

👎 Deprecated since 0.6.4:
Please use .as_mut_slice().as_mut_ptr() instead

Returns an unsafe mutable pointer to the enum map's slice.

The caller must ensure that the slice outlives the pointer this function returns, or else it will end up pointing to garbage.

Examples

use enum_map::{enum_map, EnumMap};

let mut map = enum_map! { _ => 0 };
unsafe {
    *map.as_mut_ptr().offset(11) = 23
};
assert_eq!(map[11], 23);