enum_map

Struct EnumMap

pub struct EnumMap<K: EnumArray<V>, V> { /* private fields */ }

An enum mapping.

This internally uses an array which stores a value for each possible enum value. To work, it requires implementation of internal (private, although public due to macro limitations) trait which allows extracting information about an enum, which can be automatically generated using #[derive(Enum)] macro.

Additionally, bool and u8 automatically derives from Enum. While u8 is not technically an enum, itโs convenient to consider it like one. In particular, reverse-complement in benchmark game could be using u8 as an enum.

Examples

use enum_map::{enum_map, Enum, EnumMap};

#[derive(Enum)]
enum Example {
    A,
    B,
    C,
}

let mut map = EnumMap::default();
// new initializes map with default values
assert_eq!(map[Example::A], 0);
map[Example::A] = 3;
assert_eq!(map[Example::A], 3);

Implementations

impl<K: EnumArray<V>, V> EnumMap<K, V>

pub fn values(&self) -> Values<'_, V>

An iterator visiting all values. The iterator type is &V.

Examples

use enum_map::enum_map;

let map = enum_map! { false => 3, true => 4 };
let mut values = map.values();
assert_eq!(values.next(), Some(&3));
assert_eq!(values.next(), Some(&4));
assert_eq!(values.next(), None);

pub fn values_mut(&mut self) -> ValuesMut<'_, V>

An iterator visiting all values mutably. The iterator type is &mut V.

Examples

use enum_map::enum_map;

let mut map = enum_map! { _ => 2 };
for value in map.values_mut() {
    *value += 2;
}
assert_eq!(map[false], 4);
assert_eq!(map[true], 4);

pub fn into_values(self) -> IntoValues<K, V>

Creates a consuming iterator visiting all the values. The map cannot be used after calling this. The iterator element type is V.

Examples

use enum_map::enum_map;

let mut map = enum_map! { false => "hello", true => "goodbye" };
assert_eq!(map.into_values().collect::<Vec<_>>(), ["hello", "goodbye"]);

impl<K: EnumArray<V>, V: Default> EnumMap<K, V>

pub fn clear(&mut self)

Clear enum map with default values.

Examples

use enum_map::{Enum, EnumMap};

#[derive(Enum)]
enum Example {
    A,
    B,
}

let mut enum_map = EnumMap::<_, String>::default();
enum_map[Example::B] = "foo".into();
enum_map.clear();
assert_eq!(enum_map[Example::A], "");
assert_eq!(enum_map[Example::B], "");

impl<K: EnumArray<V>, V> EnumMap<K, V>

pub const fn from_array(array: K::Array) -> EnumMap<K, V>

Creates an enum map from array.

pub fn from_fn<F>(cb: F) -> Self
where F: FnMut(K) -> V,

Create an enum map, where each value is the returned value from cb using provided enum key.

use enum_map::{enum_map, Enum, EnumMap};

#[derive(Enum, PartialEq, Debug)]
enum Example {
    A,
    B,
}

let map = EnumMap::from_fn(|k| k == Example::A);
assert_eq!(map, enum_map! { Example::A => true, Example::B => false })

pub fn iter(&self) -> Iter<'_, K, V>

Returns an iterator over enum map.

The iteration order is deterministic, and when using Enum derive it will be the order in which enum variants are declared.

Examples

use enum_map::{enum_map, Enum};

#[derive(Enum, PartialEq)]
enum E {
    A,
    B,
    C,
}

let map = enum_map! { E::A => 1, E::B => 2, E::C => 3};
assert!(map.iter().eq([(E::A, &1), (E::B, &2), (E::C, &3)]));

pub fn iter_mut(&mut self) -> IterMut<'_, K, V>

Returns a mutable iterator over enum map.

pub const fn len(&self) -> usize

Returns number of elements in enum map.

pub fn swap(&mut self, a: K, b: K)

Swaps two indexes.

Examples

use enum_map::enum_map;

let mut map = enum_map! { false => 0, true => 1 };
map.swap(false, true);
assert_eq!(map[false], 1);
assert_eq!(map[true], 0);

pub fn into_array(self) -> K::Array

Consumes an enum map and returns the underlying array.

The order of elements is deterministic, and when using Enum derive it will be the order in which enum variants are declared.

Examples

use enum_map::{enum_map, Enum};

#[derive(Enum, PartialEq)]
enum E {
    A,
    B,
    C,
}

let map = enum_map! { E::A => 1, E::B => 2, E::C => 3};
assert_eq!(map.into_array(), [1, 2, 3]);

pub const fn as_array(&self) -> &K::Array

Returns a reference to the underlying array.

The order of elements is deterministic, and when using Enum derive it will be the order in which enum variants are declared.

Examples

use enum_map::{enum_map, Enum};

#[derive(Enum, PartialEq)]
enum E {
    A,
    B,
    C,
}

let map = enum_map! { E::A => 1, E::B => 2, E::C => 3};
assert_eq!(map.as_array(), &[1, 2, 3]);

pub fn as_mut_array(&mut self) -> &mut K::Array

Returns a mutable reference to the underlying array.

The order of elements is deterministic, and when using Enum derive it will be the order in which enum variants are declared.

Examples

use enum_map::{enum_map, Enum};

#[derive(Enum, PartialEq)]
enum E {
    A,
    B,
    C,
}

let mut map = enum_map! { E::A => 1, E::B => 2, E::C => 3};
map.as_mut_array()[1] = 42;
assert_eq!(map.as_array(), &[1, 42, 3]);

pub fn as_slice(&self) -> &[V]

Converts an enum map to a slice representing values.

The order of elements is deterministic, and when using Enum derive it will be the order in which enum variants are declared.

Examples

use enum_map::{enum_map, Enum};

#[derive(Enum, PartialEq)]
enum E {
    A,
    B,
    C,
}

let map = enum_map! { E::A => 1, E::B => 2, E::C => 3};
assert_eq!(map.as_slice(), &[1, 2, 3]);

pub fn as_mut_slice(&mut self) -> &mut [V]

Converts a mutable enum map to a mutable slice representing values.

pub fn map<F, T>(self, f: F) -> EnumMap<K, T>
where F: FnMut(K, V) -> T, K: EnumArray<T>,

Returns an enum map with function f applied to each element in order.

Examples

use enum_map::enum_map;

let a = enum_map! { false => 0, true => 1 };
let b = a.map(|_, x| f64::from(x) + 0.5);
assert_eq!(b, enum_map! { false => 0.5, true => 1.5 });