enum_map
Trait EnumArray

pub trait EnumArray<V>: Enum {
    type Array: Array<V>;
}

Trait associating enum with an array.

This exists due to limitations of Rust compiler that prevent arrays from using associated constants in structures. The array length must match LENGTH of an Enum.

Required Associated Types

type Array: Array<V>

Representation of an enum map for type V.

Object Safety

This trait is not object safe.